First crack at pre-loading the attribute values.
Still need to deal with date fields.

Also note that the cascaded stuff undermines us to some extent.
Certainly when there is no value set.

Need to investigate that a bit more.

Note that this has been done by over hauling the core attribute search
so we need views for all types of attributes.